gemini-observer/requirements.txt
aiogram>=3.0.0
google-generativeai>=0.3.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
